requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
Fibo==1.0.0
idna==3.10
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
